requirements.txt
mage-ai
google-cloud-bigquery
pandas-gbq
db-dtypes
praw
python-dotenv
transformers
torch --index-url https://download.pytorch.org/whl/cu118
bertopic
sentence-transformers
langdetect
spacy
nltk
matplotlib
seaborn
scikit-learn
nltk
plotly
kaleido
sidrapy